Delete request coded.

+++ Front-end/controller/HomeController.js
@@ -5,7 +5,7 @@ $('#btnSearchOrder').click(function () {
         let searchOrder = $('#searchOrder').val();
         let chooseOrderType = $('#chooseOrderType').val();
         if (chooseOrderType === "ID") {
-            console.log("ID : "+searchOrder +"==="+ search.orId)
+            console.log("ID : " + searchOrder + "===" + search.orId)
 
             if (searchOrder === search.orId) {
                 $('#orderIdDash').val(search.orId);
@@ -16,7 +16,7 @@ $('#btnSearchOrder').click(function () {
 
             }
         } else if (chooseOrderType === "1") {
-            console.log("1 : "+searchOrder +"==="+ search.cusName)
+            console.log("1 : " + searchOrder + "===" + search.cusName)
             if (searchOrder === search.orCusName) {
                 $('#orderIdDash').val(search.orId);
                 $('#OrderDateDash').val(search.orDate);
@@ -25,7 +25,7 @@ $('#btnSearchOrder').click(function () {
                 $('#subTotDash').val(search.orSubTotal);
             }
         } else if (chooseOrderType === "2") {
-            console.log("2 : "+searchOrder +"==="+ search.ordDate)
+            console.log("2 : " + searchOrder + "===" + search.ordDate)
 
             if (searchOrder === search.orDate) {
                 $('#orderIdDash').val(search.orId);
@@ -39,7 +39,7 @@ $('#btnSearchOrder').click(function () {
     }
 });
 
-$('#btnClearOrd').click(function (){
+$('#btnClearOrd').click(function () {
     $('#orderIdDash').val("");
     $('#OrderDateDash').val("");
     $('#customerNameDash').val("");
@@ -48,43 +48,83 @@ $('#btnClearOrd').click(function (){
     $('#searchOrder').val("");
 });
 
-$('#btnDeleteOrd').click(function (){
-    let deleteOrderId = $('#orderIdDash').val();
+$('#btnDeleteOrd').click(function () {
+    Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            let deleteOrderId = $('#orderIdDash').val();
+            console.log($("#orderIdDash"));
+            if (deleteOrder(deleteOrderId)) {
+                setOrderTextfieldValues("", "", "", "", "");
+                Swal.fire({
+                    position: 'top-end',
+                    icon: 'success',
+                    title: 'Your file has been deleted.!',
+                    showConfirmButton: false,
+                    timer: 1500
+                });
 
-    if (deleteOrder(deleteOrderId)){
-        alert("Order Successfully Deleted....");
-        setOrderTextfieldValues("", "", "", "","");
-    }else{
-        alert("No such Order to delete. please check the id");
-    }
+            } else {
+                Swal.fire({
+                    position: 'top-end',
+                    icon: 'warning',
+                    title: 'No such Order to delete. please check the id !',
+                    showConfirmButton: false,
+                    timer: 1500
+                });
+            }
+        }
+    })
+
 });
 
 
 /*FUNCTIONS*/
 
-function searchOrder(orderId){
-    for(var i of orders){
-        if (i.orId === orderId){
+function searchOrder(orderId) {
+    for (let i of orders) {
+        if (i.orId === orderId) {
             return i;
         }
     }
     return null;
 }
 
-function deleteOrder(orderId){
+function deleteOrderRequest(orderId) {
+    $.ajax({
+        url: "http://localhost:8080/Back_end_war/placeorder?orderId=".concat(orderId),
+        method: "DELETE",
+        success:function (resp) {
+            console.log(resp);
+            let indexNumber = orders.indexOf(orderId);
+            orders.splice(indexNumber, 1);
+            loadAllOrder();
+        },
+        error:function (err) {
+
+        }
+    });
+}
+
+function deleteOrder(orderId) {
     let ordObj = searchOrder(orderId);
 
-    if (ordObj != null){
-        let indexNumber = orders.indexOf(ordObj);
-        orders.splice(indexNumber,1);
-        loadAllOrder();
+    if (ordObj != null) {
+        deleteOrderRequest(orderId);
         return true;
-    }else {
+    } else {
         return false;
     }
 }
 
-function setOrderTextfieldValues(orderId,date,name,dis,cost){
+function setOrderTextfieldValues(orderId, date, name, dis, cost) {
 
     $('#orderIdDash').val(orderId);
     $('#OrderDateDash').val(date);
